Extract server part of notifier.
