+++ coveralls.json
@@ -3,7 +3,8 @@
     "minimum_coverage": 87
   },
   "skip_files": [
-    "test/support"
+    "test/support",
+    "lib/course_planner/notifier.ex"
   ]
 }
