+++ test/lib/course_planner/notifier/server_test.exs
@@ -1,11 +1,11 @@
-defmodule CoursePlanner.NotifierTest do
+defmodule CoursePlanner.Notifier.ServerTest do
   use CoursePlanner.ModelCase
-  doctest CoursePlanner.Notifier
+  doctest CoursePlanner.Notifier.Server
 
   import CoursePlanner.Factory
   import Swoosh.TestAssertions
 
-  alias CoursePlanner.{Notifier, Notifications, User}
+  alias CoursePlanner.{Notifier.Server, Notifications, User}
 
   test "save notification to send later" do
     user = insert(:user)
@@ -14,7 +14,7 @@ defmodule CoursePlanner.NotifierTest do
     |> Notifications.resource_path("/")
     |> Notifications.to(user)
 
-    Notifier.handle_cast({:save_email, notification}, [])
+    Server.handle_cast({:save_email, notification}, [])
     saved_user = Repo.get(User, user.id) |> Repo.preload(:notifications)
     [saved_notification] = saved_user.notifications
     assert saved_notification.type == "user_modified"
@@ -28,7 +28,7 @@ defmodule CoursePlanner.NotifierTest do
 
     saved_user = Repo.get(User, user.id) |> Repo.preload(:notifications)
 
-    Notifier.handle_cast({:notify_all, saved_user}, [])
+    Server.handle_cast({:notify_all, saved_user}, [])
     assert_email_sent subject: "Activity Summary"
     sent_user = Repo.get(User, user.id) |> Repo.preload(:notifications)
     assert sent_user.notifications == []
@@ -42,7 +42,7 @@ defmodule CoursePlanner.NotifierTest do
     |> Notifications.resource_path("/")
     |> Notifications.to(user)
 
-    Notifier.handle_cast({:send_email, notification}, [])
+    Server.handle_cast({:send_email, notification}, [])
     assert_email_sent subject: "Your profile is updated"
   end
